fix refresh
Because of wtimeout(dsp_viewport, n), key input needs to be
wgetch(dsp_viewport) to keep things moving around. Also, it allows for
"pausing" when in a store so the windows don't keep updating as you try
to buy/sell items. This still happens when selecting a speed though,
which needs to be fixed.